esp32, drivers, lib: Change LoRa timebase from 1000us to 1ms.

+++ esp32/lora/timer-board.h
@@ -34,7 +34,7 @@ Maintainer: Miguel Luis and Gregory Cristian
  * \brief Timer time variable definition
  */
 #ifndef TimerTime_t
-typedef uint64_t TimerTime_t;
+typedef uint32_t TimerTime_t;
 #endif
 
 /*!
